update sqlc and feature

+++ common/postgres/migration/000001_init.up.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS users (
 CREATE TABLE IF NOT EXISTS products (
   id uuid PRIMARY KEY NOT NULL DEFAULT gen_random_uuid(),
   name varchar(255) NOT NULL,
-  price numeric(10,2),
+  price integer,
   stock integer NOT NULL,
   is_service boolean NOT NULL DEFAULT false,
   description varchar(255) NOT NULL,
